Moved UsersTable to components directory and added AdminUsersPage.jsx
